pyproject.toml
[project]
name = "domestic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.16.1",
    "dominate>=2.9.1",
    "fastapi[all]>=0.115.12",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.24.0",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
    "python-jose[cryptography]>=3.3.0",
    "requests>=2.28.1",
    "pyjwt>=2.10.1",
    "jinja2>=3.1.6",
    "pydantic>=2.11.5",
]

[tool.pytest]
asyncio_mode = "auto"
pythonpath = ["."]

#[tool.uv]
#package = true

[tool.pytest.ini_options]
pythonpath = [
    "src"
]
